Lesson 16.1
Writing the function to move the player.

Note; there were some invalid references.  The code as is (copied and pasted) does not run because the button click is now appended with a _1 value, for example private void btnNorth_Click_1 instead of private void btnNorth_Click.  We can clean this code up later.